Harden transformers + thread list-loaded limit
