+++ src/JKToolKit.CodexSDK.Demo/Commands/AppServerThreads/AppServerThreadListLoadedCommand.cs
@@ -12,10 +12,11 @@ public sealed class AppServerThreadListLoadedCommand : AsyncCommand<AppServerThr
         CancellationToken cancellationToken) =>
         AppServerThreadCommandHelpers.RunWithClientAsync(settings, cancellationToken, async (codex, ct) =>
         {
+            var limit = settings.Limit > 0 ? settings.Limit : null;
             var page = await codex.ListLoadedThreadsAsync(new ThreadLoadedListOptions
             {
                 Cursor = string.IsNullOrWhiteSpace(settings.Cursor) ? null : settings.Cursor,
-                Limit = settings.Limit
+                Limit = limit
             }, ct);
 
             if (settings.Json)
